게임 생성 라우팅으로 변경
