+++ src/index.tsx
@@ -11,6 +11,8 @@ import reportWebVitals from "./reportWebVitals";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import GlobalStyle from "./styles/global-styles";
 import { theme } from "./styles/theme";
+import { SelectGolfCourse } from "./pages/MakeGame/SelectGolfCourse/SelectGolfCourse";
+import Setup from "./pages/MakeGame/Setup";
 
 // TODO : 현재 도메인이 /pwa-react-test라 반드시 붙여야하는지? 확인 필요
 const router = createBrowserRouter([
@@ -27,6 +29,16 @@ const router = createBrowserRouter([
   {
     path: "/make_game",
     element: <MakeGame />,
+    children: [
+      {
+        path: "",
+        element: <Setup />,
+      },
+      {
+        path: "select_golf_course",
+        element: <SelectGolfCourse />,
+      },
+    ],
     errorElement: <ErrorPage />,
   },
 ]);
