+++ src/pages/MakeGame/SelectGolfCourse/index.ts
@@ -1,0 +1,3 @@
+import { SelectGolfCourse } from "./SelectGolfCourse";
+
+export default SelectGolfCourse;
