+++ src/pages/MakeGame/Setup/index.ts
@@ -1,0 +1,3 @@
+import { Setup } from "./Setup";
+
+export default Setup;
